picture api modefied

+++ controller/picture.js
@@ -16,17 +16,34 @@ class PictureController {
 
   /**
    *
-   * delete and the create
+   * delete first and then create
    * @param {*} ctx
    * @returns data
    * @memberof PictureController
    */
   async create(ctx) {
     const { fileList, category } = ctx.request.body
-    await Picture.destroy({
-      where: { category }
+    if(Array.isArray(fileList)) {
+      await Picture.destroy({
+        where: { category }
+      })
+      return Picture.bulkCreate(fileList)
+    } else {
+      return Picture.create(fileList)
+    }
+  }
+  
+  async update(ctx) {
+    const { body } = ctx.request
+    console.log(body)
+    return Picture.update({ 
+      title: body.title,
+      url: body.url,
+      name: body.name,
+      uid: body.uid
+    }, {
+      where: { id: body.id }
     })
-    return Picture.bulkCreate(fileList)
   }
 }
